requirements.txt
absl-py==0.9.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
google-auth==1.13.1
google-auth-oauthlib==0.4.1
grpcio==1.28.1
idna==2.9
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.1
munch==2.5.0
numpy==1.18.2
oauthlib==3.1.0
Pillow==7.1.1
pretrainedmodels==0.7.4
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
sentencepiece==0.1.85
six==1.14.0
tensorboard==2.2.0
tensorboard-plugin-wit==1.6.0.post3
torch==1.4.0
torchtext==0.5.0
torchvision==0.5.0
tqdm==4.45.0
urllib3==1.25.8
Werkzeug==1.0.1
